Change handling of missing data in to_array_vlf method
The implementation of `to_array_vlf` is problematic since it skips
missing keys in any of the dimensions and it is not possible to derive
for which key data was missing. For instance, if not all requested
validity times are available, the method will only return the available
ones and the user does not know for which validity times the data was
missing.

This changes the behavior of the method such that the shape always
matches the requested size and all entries where data is missing are set
to `NaN`.

Signed-off-by: cwasicki <126617870+cwasicki@users.noreply.github.com>